requirements.txt
wheel
djangorestframework
django-filter
# swagger
drf-yasg
djangorestframework-simplejwt
python-decouple
Pillow
psycopg2-binary
python-slugify
celery
Jinja2
redis